% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thing.R
\name{thing}
\alias{thing}
\title{A function messaging a thing}
\usage{
thing()
}
\value{
Nothing, print thing
}
\description{
A function messaging a thing
}
\examples{
thing()
}
\section{Custom section}{
\itemize{
  \item Hello I am here
}
}

